+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Add `sdwan_transport_ipv6_tracker_group_profile_parcel` resource and data source
 - Add `sdwan_service_object_tracker_profile_parcel` resource and data source
 - Add `sdwan_service_object_tracker_group_profile_parcel` resource and data source
-- Add `sdwan_service_routing_bgp_profile_parcel` resource and data source
+- Add `sdwan_service_bgp_routing_profile_parcel` resource and data source
 - Add `sdwan_service_switchport_profile_parcel` resource and data source
 
 ## 0.3.13